requirements.txt
kaggle
kafka-python
pyspark==3.3.0
pandas
streamlit
plotly
pyarrow
findspark
